fix install magento replacing composer.json

+++ build-packages/magento-scripts/lib/tasks/magento/install-magento.js
@@ -5,6 +5,8 @@ const fs = require('fs');
 const runComposerCommand = require('../../util/run-composer');
 const matchFilesystem = require('../../util/match-filesystem');
 const moveFile = require('../../util/move-file');
+const pathExists = require('../../util/path-exists');
+const getJsonFileData = require('../../util/get-jsonfile-data');
 
 const installMagento = {
     title: 'Installing Magento',
@@ -25,21 +27,59 @@ const installMagento = {
 
         task.title = 'Creating Magento project...';
 
-        const tempDir = path.join(os.tmpdir(), `magento-tmpdir-${Date.now()}`);
-        await runComposerCommand(
-            `create-project \
-        --repository=https://repo.magento.com/ magento/project-community-edition=${magentoVersion} \
-        --no-install \
-        "${tempDir}"`,
-            { magentoVersion }
-        );
+        if (await pathExists(path.join(config.magentoDir, 'composer.json'))) {
+            const composerData = await getJsonFileData(path.join(config.magentoDir, 'composer.json'));
 
-        await moveFile({
-            from: path.join(tempDir, 'composer.json'),
-            to: path.join(process.cwd(), 'composer.json')
-        });
+            if (!composerData.repositories
+                || (Array.isArray(composerData.repositories)
+                    && !composerData.repositories.some((repo) => repo.type === 'composer' && repo.url === 'https://repo.magento.com/')
+                )
+                || (typeof composerData.repositories === 'object'
+                    && !Object.values(composerData.repositories).some((repo) => repo.type === 'composer' && repo.url === 'https://repo.magento.com/'))
+            ) {
+                await runComposerCommand('config repo.0 composer https://repo.magento.com', {
+                    magentoVersion,
+                    callback: (t) => {
+                        task.output = t;
+                    }
+                });
+            }
 
-        await fs.promises.rmdir(tempDir);
+            if (!composerData.require['magento/product-community-edition']) {
+                await runComposerCommand(`require magento/product-community-edition:${magentoVersion}`,
+                    {
+                        magentoVersion,
+                        callback: (t) => {
+                            task.output = t;
+                        }
+                    });
+            }
+            if (!composerData.require['magento/composer-root-update-plugin']) {
+                await runComposerCommand('require magento/composer-root-update-plugin:~1.0',
+                    {
+                        magentoVersion,
+                        callback: (t) => {
+                            task.output = t;
+                        }
+                    });
+            }
+        } else {
+            const tempDir = path.join(os.tmpdir(), `magento-tmpdir-${Date.now()}`);
+            await runComposerCommand(
+                `create-project \
+            --repository=https://repo.magento.com/ magento/project-community-edition=${magentoVersion} \
+            --no-install \
+            "${tempDir}"`,
+                { magentoVersion }
+            );
+
+            await moveFile({
+                from: path.join(tempDir, 'composer.json'),
+                to: path.join(process.cwd(), 'composer.json')
+            });
+
+            await fs.promises.rmdir(tempDir);
+        }
 
         try {
             await runComposerCommand('install',
